move genkeys and sandboxdatabase into test _init
